fixed scandir def to compile with both new and old versions added step in iocommon_readorwrite to generate PVFS_Request
SVN revision 9106

+++ src/client/usrint/posix-ops.h
@@ -104,14 +104,14 @@ typedef struct posix_ops_s
                      const void *value, size_t size, int flags);
     int (*fsetxattr)(int fd, const char *name,
                      const void *value, size_t size, int flags);
-    int (*getxattr)(const char *path, const char *name,
-                    void *value, size_t size);
-    int (*lgetxattr)(const char *path, const char *name,
-                     void *value, size_t size);
-    int (*fgetxattr)(int fd, const char *name, void *value, size_t size);
-    int (*listxattr)(const char *path, char *list, size_t size);
-    int (*llistxattr)(const char *path, char *list, size_t size);
-    int (*flistxattr)(int fd, char *list, size_t size);
+    ssize_t (*getxattr)(const char *path, const char *name,
+                        void *value, size_t size);
+    ssize_t (*lgetxattr)(const char *path, const char *name,
+                         void *value, size_t size);
+    ssize_t (*fgetxattr)(int fd, const char *name, void *value, size_t size);
+    ssize_t (*listxattr)(const char *path, char *list, size_t size);
+    ssize_t (*llistxattr)(const char *path, char *list, size_t size);
+    ssize_t (*flistxattr)(int fd, char *list, size_t size);
     int (*removexattr)(const char *path, const char *name);
     int (*lremovexattr)(const char *path, const char *name);
     int (*fremovexattr)(int fd, const char *name);
